src/commands/generateApi.ts
import { Command } from 'commander'
import * as fs from 'fs-extra'
import * as path from 'path'

const generateApi = new Command()

generateApi
	.command('generate-api')
	.description('Generate an AWS API endpoint with optional features')
	.argument('<name>', 'Name of the API')
	.option('-a, --auth', 'Include AWS Cognito authentication')
	.option('-d, --db', 'Integrate with PostgreSQL database')
	.action((name, options) => {
		let apiTemplate = `export const ${name} = async (event) => {
      // Your API logic here`

		if (options.auth) {
			apiTemplate += `
      // Authenticate with AWS Cognito
      const user = await authenticateUser(event);`
		}

		if (options.db) {
			apiTemplate += `
      // Database interaction with PostgreSQL
      const data = await fetchDataFromDB();`
		}

		apiTemplate += `
      return {
        statusCode: 200,
        body: JSON.stringify({ message: "${name} endpoint generated" }),
      };
    };`

		if (options.auth) {
			apiTemplate += `

    // Function to authenticate user with AWS Cognito
    const authenticateUser = async (event) => {
      // Implement Cognito authentication logic here
    };`
		}

		if (options.db) {
			apiTemplate += `

    // Function to interact with PostgreSQL database
    const fetchDataFromDB = async () => {
      // Implement database logic here
    };`
		}

		const outputPath = path.resolve(__dirname, `../output/${name}.ts`)
		fs.outputFileSync(outputPath, apiTemplate)

		console.log(`API endpoint ${name} generated at ${outputPath}`)
	})

export default generateApi
